updates 20160514

+++ app/src/main/java/com/jcaseydev/popularmovies/Backend/Movie.java
@@ -45,7 +45,7 @@ public class Movie implements Parcelable{
         return movieVoteAverage;
     }
 
-    public double getMovieId(){
+    public int getMovieId(){
         return movieId;
     }
